Move New Keynesian stuff to new file
Complicated methods that are used *only* by HANK applications have been moved to a new file. The SSJ example notebook has been updated to account for the changes, as have the unit tests.
