+++ HARK/ConsumptionSaving/tests/test_IndShockConsumerType.py
@@ -7,7 +7,6 @@ from HARK.ConsumptionSaving.ConsIndShockModel import (
     IndShockConsumerType,
     init_idiosyncratic_shocks,
     init_lifecycle,
-    additional_HANK_parameters,
 )
 from HARK.tests import HARK_PRECISION
 
@@ -638,9 +637,6 @@ dict_harmenberg = {
     "mFac": 3,
 }
 
-jacobian_test_dict = dict_harmenberg.copy()
-jacobian_test_dict.update(additional_HANK_parameters)
-
 
 class test_Harmenbergs_method(unittest.TestCase):
     def test_Harmenberg_mtd(self):
@@ -840,52 +836,3 @@ class testLCMortalityReadShocks(unittest.TestCase):
         # (the exception from before should not happen
         # because we are killing agents before T_cycle)
         self.assertTrue(np.all(hist["t_age"] == hist["t_cycle"]))
-
-
-# %% Test Transition Matrix Methods
-
-
-class test_Transition_Matrix_Methods(unittest.TestCase):
-    def test_calc_tran_matrix(self):
-        example1 = IndShockConsumerType(**dict_harmenberg)
-        example1.cycles = 0
-        example1.solve()
-
-        example1.define_distribution_grid()
-        p = example1.dist_pGrid  # Grid of permanent income levels
-
-        example1.calc_transition_matrix()
-        c = example1.cPol_Grid  # Normalized Consumption Policy Grid
-        asset = example1.aPol_Grid  # Normalized Asset Policy Grid
-
-        example1.calc_ergodic_dist()
-        vecDstn = example1.vec_erg_dstn
-        # Distribution of market resources and permanent income as a vector (m*p)x1 vector where
-
-        # Compute Aggregate Consumption and Aggregate Assets
-        gridc = np.zeros((len(c), len(p)))
-        grida = np.zeros((len(asset), len(p)))
-
-        for j in range(len(p)):
-            gridc[:, j] = p[j] * c  # unnormalized Consumption policy grid
-            grida[:, j] = p[j] * asset  # unnormalized Asset policy grid
-
-        AggC = np.dot(gridc.flatten(), vecDstn)  # Aggregate Consumption
-        AggA = np.dot(grida.flatten(), vecDstn)  # Aggregate Assets
-
-        self.assertAlmostEqual(AggA[0], 1.19513, places=4)
-        self.assertAlmostEqual(AggC[0], 1.00417, places=4)
-
-
-# %% Test Heterogenous Agent Jacobian Methods
-
-
-class test_Jacobian_methods(unittest.TestCase):
-    def test_calc_jacobian(self):
-        Agent = IndShockConsumerType(**jacobian_test_dict)
-        Agent.compute_steady_state()
-        CJAC_Perm, AJAC_Perm = Agent.calc_jacobian("PermShkStd", 50)
-
-        self.assertAlmostEqual(CJAC_Perm.T[30][29], -0.06120, places=HARK_PRECISION)
-        self.assertAlmostEqual(CJAC_Perm.T[30][30], 0.05307, places=HARK_PRECISION)
-        self.assertAlmostEqual(CJAC_Perm.T[30][31], 0.04674, places=HARK_PRECISION)
